Improve genetic path optimizer with the help of https://github.com/CodingTrain/website/blob/master/CodingChallenges/CC_035.4_TSP_GA/Processing/

+++ prolab/2/1/src/Route.java
@@ -5,6 +5,6 @@ import java.util.ArrayList;
  */
 public class Route
 {
-	public double cost = 0;
+	public float cost = 0;
 	public ArrayList<City> cities = new ArrayList<>();
 }
